# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统API接口文档",
        "title": "通用后台管理系统",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/admin/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新增用户接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "新增用户接口",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AddSysAdminDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/captcha": {
            "get": {
                "description": "验证码接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证码"
                ],
                "summary": "验证码接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "用户登录接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录接口",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.LoginDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新增岗位",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "新增岗位",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AddSysPostDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/batchDel": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "批量删除岗位",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "批量删除岗位",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.BatchDelSysPostDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/changeStatus": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "改变岗位状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "改变岗位状态",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateSysPostStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/del/{postId}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除岗位",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "删除岗位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "岗位ID",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/info/{postId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取岗位详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "获取岗位详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "岗位ID",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取岗位列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "岗位列表(分页)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "岗位名称",
                        "name": "postName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "岗位状态:1启用,2禁用",
                        "name": "postStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "beginTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "endTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/select": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取岗位列表(下拉框)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "获取岗位列表(下拉框)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/post/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "编辑岗位",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位管理"
                ],
                "summary": "编辑岗位",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateSysPostDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.AddSysAdminDto": {
            "type": "object",
            "required": [
                "deptId",
                "email",
                "nickname",
                "password",
                "phone",
                "postId",
                "rostId",
                "status",
                "username"
            ],
            "properties": {
                "deptId": {
                    "description": "部门id",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "note": {
                    "description": "备注",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "postId": {
                    "description": "岗位id",
                    "type": "integer"
                },
                "rostId": {
                    "description": "角色id",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1-\u003e启用,2-\u003e禁用",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "entity.AddSysPostDto": {
            "type": "object",
            "properties": {
                "postCode": {
                    "description": "岗位编码",
                    "type": "string"
                },
                "postName": {
                    "description": "岗位名称",
                    "type": "string"
                },
                "postStatus": {
                    "description": "岗位状态(1:正常 2:停用)",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                }
            }
        },
        "entity.BatchDelSysPostDto": {
            "type": "object",
            "properties": {
                "ids": {
                    "description": "岗位ID数组",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "entity.LoginDto": {
            "type": "object",
            "required": [
                "captcha",
                "idKey",
                "password",
                "username"
            ],
            "properties": {
                "captcha": {
                    "description": "验证码",
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 4
                },
                "idKey": {
                    "description": "验证码key",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "entity.UpdateSysPostDto": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "岗位ID",
                    "type": "integer"
                },
                "postCode": {
                    "description": "岗位编码",
                    "type": "string"
                },
                "postName": {
                    "description": "岗位名称",
                    "type": "string"
                },
                "postStatus": {
                    "description": "岗位状态(1:正常 2:停用)",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                }
            }
        },
        "entity.UpdateSysPostStatusDto": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "岗位ID",
                    "type": "integer"
                },
                "postStatus": {
                    "description": "岗位状态(1:正常 2:停用)",
                    "type": "integer"
                }
            }
        },
        "response.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}